Refactor shared testing database implementation
I implemented a sessionmaker, so that in future, the test database could persist over multiple sessions for testing/development purposes. It seemed like a cleaner approach. I also renamed the functions in tests\utils\db.py to be explicit that they were in-memory databases, in case I add another test database setup in the future.