Refactor testing to use shared movie_data fixture
Since the enum for rating couldn't be passed via json, I removed the rating field from movie_data and added it back by hand to tests that required it. This speaks to a larger problem with the implementation of "rating" and how it can be set via the route, but that's a problem for another day.